Intercept search algorithm to fetch agenda
Each time we move the agent to the last part of the path in the agenda,
this is a crap visualisation, but demonstrates the pipeline works. We
know need to figure out how to handle backtracking behaviour.

Currently the agenda items only have the h/g costs, and path, however,
for a search algorithm using backtracking, we will often change tact and
select an agenda item that is not an extension of the current agenda
item's path. This results in confusing animation: "What is the agent
doing?". Instead we need to graphically demonstrate the backtracking
through animation